requirements.txt
Flask
msal
pyodbc
azure-identity